stubs for linux added
The name on `DllImport` is `cairo`, therefore the runtime evaluates the dll to
* cairo.dll on windows
* libcairo.so on linux

But on linux the actual so is libcairo.so.2 (see /usr/lib/x86_64-linux-gnu, there it is a symlink to libcairo.so.2.11400.6).
Therefore the so can't be loaded. Possible workaurounds are:
* another symlink ln -s libcairo.so.2 libcairo.so
* a stub libcairo.so that loads libcairo.so.2

The first approach is possible, but requires user-action (and sudo) therefore this is error-prone and not practical.
The second approach requires no user action, hence this one is preferable.